Modularise runtime integration tests
